requirements.txt
opencv-python-headless
face_recognition
streamlit
yagmail
numpy
Pillow
